requirements.txt
Telethon==1.33.1
python-dotenv==1.0.1
